requirements.txt
setuptools
wheel
python-dotenv
requests
